feat: add validator + fix models

+++ handlers/permissionHandler.js
@@ -1,6 +1,13 @@
+const { validationResult } = require('express-validator');
 const models = require('../models');
 
 async function createPermission(req, res) {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
   try {
     const { name } = req.body;
     const newPermission = await models.Permissions.create({ name });
@@ -26,6 +33,12 @@ async function getAllPermissions(req, res) {
 }
 
 async function updatePermission(req, res) {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
   try {
     const permissionId = req.params.id;
     const { name } = req.body;
@@ -42,6 +55,12 @@ async function updatePermission(req, res) {
 }
 
 async function deletePermission(req, res) {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
   try {
     const permissionId = req.params.id;
     await models.Permissions.destroy({ where: { id: permissionId } });
